+++ tests/testthat/test_applyChangeMeta.R
@@ -138,7 +138,7 @@ test_that("Changes to all_GADSdat on value level", {
 })
 
 
-test_that("Changes to GADSdat if tibble, varNames", {
+test_that("Changes to GADSdat if tibble or data.frame, varNames", {
   # varName
   changes_var_tbl <- changes_var
   changes_var_tbl[1, "varName_new"] <- "new1"
@@ -149,9 +149,13 @@ test_that("Changes to GADSdat if tibble, varNames", {
   expect_equal(g1$labels$varName, c(rep("new1", 3), rep("VAR2", 2), rep("VAR3", 2)))
   expect_equal(names(g1$dat), c("new1", "VAR2", "VAR3"))
 
+  changes_var_df <- as.data.frame(changes_var_tbl)
+  g3 <- applyChangeMeta(changes_var_df, dfSAV)
+  expect_equal(g3$labels[, -1], dfSAV$labels[, -1])
+
 })
 
-test_that("Changes to GADSdat if tibble, value labels", {
+test_that("Changes to GADSdat if tibble or data.frame, value labels", {
   changes_val_tbl <- changes_val
   changes_val_tbl[1, "valLabel_new"] <- "new_miss"
   changes_val_tbl[2, "valLabel_new"] <- "new_miss2"
@@ -161,4 +165,8 @@ test_that("Changes to GADSdat if tibble, value labels", {
   expect_equal(g2$labels[, -7], dfSAV$labels[, -7])
   expect_equal(g2$labels$valLabel, c("new_miss", "new_miss2", "One", "missing", NA, "missing", NA))
   expect_equal(names(g2$dat), names(dfSAV$dat))
+
+  changes_val_df <- as.data.frame(changes_val_tbl)
+  g4 <- applyChangeMeta(changes_val_df, dfSAV)
+  expect_equal(g4$labels[, -7], dfSAV$labels[, -7])
 })